+++ GoogleSecrets/EnvironmentVariableNames.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.Extensions.Configuration.GoogleSecrets
+{
+    /// <summary>
+    /// Contains the names of environment variables used for Google Secrets configuration.
+    /// </summary>
+    public static class EnvironmentVariableNames
+    {
+        /// <summary>
+        /// The name of the environment variable where the google secrets project id is stored.
+        /// </summary>
+        public const string GoogleSecretsProject = "GOOGLE_SECRETS_PROJECT";
+    }
+}